requirements.txt
pdfplumber
streamlit
pillow
ultralyticsplus==0.0.28 
ultralytics==8.0.43
natsort
pandas
paddlepaddle 
paddleocr
ghostscript
pdf2image
